CartSummary 컴포넌트 및 scss 분리 완료, CartMockData.json 생성
